Fix alignement of forced platforms on patch versions
This commit fixes the behavior of alignement whenever the
platform version is a patch version, and that not all modules
are available in that patch version. Previously, alignement
would be defeated. With this change, we will do our best to
align to the highest seen version in the graph for each of
the modules.

Fixes nebula-plugins/gradle-nebula-integration#51